requirements.txt
dnspython==2.6.1
python-dotenv==1.0.1 